R-plotmo: Autospec creation for update from version 3.3.7 to version 3.4.0
--- 3.4.0  May 31, 2018

If predict.rpart is predicting a probability, plotmo now recognizes
that and sets ylim=c(0,1) appropriately.

Plotting of intercept-only models was slightly inconsistent. Fixed that.

We now attempt to better set the default nticks in persp plots.

We now position the labels in persp plots slightly better along
the axes (they were sometimes too far away from the front corner).

When degree2 is exactly two strings, we now assign the x1 and x2 axes
in the order specified in degree2 (although persp plots still get
rotated for optimum visibility of the surface, and this rotation can

(NEWS truncated at 15 lines)